fix: place util methods into a new file
